backend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
django-cors-headers>=4.3.1
djangorestframework>=3.14.0
razorpay>=1.4.2
dj-database-url>=2.1.0
